Support sticky sessions
This allows students to move to the next section of a self paced course
without losing their container.

When they first go to the page, selfpaced.js checks if there is a
cookie with a uuid for that page. If it finds one it opens the iframe
with that uuid as a parameter, otherwise it generates a uuid and sets a
cookie.

This also adds a cron job and changes the cleanup task to clean up all
nodes that have run past the timeout.

There is also a change to the Dockerfile to improve startup times for the
containers.